move component to page
